requirements.txt
alabaster==0.7.12
asgiref==3.3.1
automium==0.2.6
automium-web==0.1.1
Babel==2.9.0
bootstrap-py==1.0.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
defusedxml==0.7.0
dj-database-url==0.5.0
Django==3.1.7
django-appconf==1.0.4
django-crispy-forms==1.11.1
django-heroku==0.3.1
django-imagekit==4.0.2
djangorestframework==3.12.3
djangorestframework-simplejwt==4.6.0
docutils==0.16
Flask==1.1.2
gitdb==4.0.5
GitPython==3.1.14
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
imagesize==1.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
legacy==0.1.6
license==0.1a3
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
pguard==0.4.0
pilkit==2.0
Pillow==8.1.1
psycopg2==2.8.6
pycparser==2.20
Pygments==2.8.0
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==1.5
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
smmap==3.0.5
snowballstemmer==2.1.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
Sphinx==3.5.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
urllib3==1.26.3
views==0.3
Werkzeug==1.0.1
whitenoise==5.2.0
